+++ caikit/runtime/servicers/model_runtime_servicer.py
@@ -24,7 +24,7 @@ from caikit.runtime.model_management.model_manager import ModelManager
 from caikit.runtime.protobufs import model_runtime_pb2, model_runtime_pb2_grpc
 from caikit.runtime.types.aborted_exception import AbortedException
 from caikit.runtime.types.caikit_runtime_exception import CaikitRuntimeException
-from caikit.runtime.work_management.abortable_action import (
+from caikit.runtime.work_management.abortable_context import (
     AbortableContext,
     ThreadInterrupter,
 )
