+++ caikit/runtime/work_management/rpc_aborter.py
@@ -25,7 +25,7 @@ import grpc
 import alog
 
 # Local
-from caikit.runtime.work_management.abortable_action import (
+from caikit.runtime.work_management.abortable_context import (
     AbortableContextBase,
     ActionAborter,
 )
